customer_support_env/server/requirements.txt
openenv-core[core]>=0.2.1
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.20.0
